Move fixtures from .mock file to utils and export as static.
