+++ src/legacy/core_plugins/data/public/index.ts
@@ -63,4 +63,6 @@ export {
   IndexPatternMissingIndices,
   NoDefaultIndexPattern,
   NoDefinedIndexPatterns,
+  mockFields,
+  mockIndexPattern,
 } from './index_patterns';
