+++ src/legacy/core_plugins/data/public/index_patterns/index.ts
@@ -18,5 +18,3 @@
  */
 
 export * from './index_patterns_service';
-
-export { mockFields, mockIndexPattern } from './index_patterns_service.mock';
