+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns_service.mock.ts
@@ -63,69 +63,3 @@ export const indexPatternsServiceMock = {
   createSetupContract: createSetupContractMock,
   createStartContract: createSetupContractMock,
 };
-
-export const mockFields: FieldType[] = [
-  {
-    name: 'machine.os',
-    esTypes: ['text'],
-    type: 'string',
-    aggregatable: false,
-    searchable: false,
-    filterable: true,
-  },
-  {
-    name: 'machine.os.raw',
-    type: 'string',
-    esTypes: ['keyword'],
-    aggregatable: true,
-    searchable: true,
-    filterable: true,
-  },
-  {
-    name: 'not.filterable',
-    type: 'string',
-    esTypes: ['text'],
-    aggregatable: true,
-    searchable: false,
-    filterable: false,
-  },
-  {
-    name: 'bytes',
-    type: 'number',
-    esTypes: ['long'],
-    aggregatable: true,
-    searchable: true,
-    filterable: true,
-  },
-  {
-    name: '@timestamp',
-    type: 'date',
-    esTypes: ['date'],
-    aggregatable: true,
-    searchable: true,
-    filterable: true,
-  },
-  {
-    name: 'clientip',
-    type: 'ip',
-    esTypes: ['ip'],
-    aggregatable: true,
-    searchable: true,
-    filterable: true,
-  },
-  {
-    name: 'bool.field',
-    type: 'boolean',
-    esTypes: ['boolean'],
-    aggregatable: true,
-    searchable: true,
-    filterable: true,
-  },
-];
-
-export const mockIndexPattern: StaticIndexPattern = {
-  id: 'logstash-*',
-  fields: mockFields,
-  title: 'logstash-*',
-  timeFieldName: '@timestamp',
-};
