+++ src/legacy/core_plugins/data/public/index_patterns/utils.ts
@@ -21,7 +21,8 @@
 import { get } from 'lodash';
 // @ts-ignore
 import { KBN_FIELD_TYPES } from '../../../../utils/kbn_field_types';
-import { Field } from './fields';
+import { Field, FieldType } from './fields';
+import { StaticIndexPattern } from './index_patterns';
 
 export const ILLEGAL_CHARACTERS = 'ILLEGAL_CHARACTERS';
 export const CONTAINS_SPACES = 'CONTAINS_SPACES';
@@ -97,3 +98,69 @@ export function getRoutes() {
     sourceFilters: '/management/kibana/index_patterns/{{id}}?_a=(tab:sourceFilters)',
   };
 }
+
+export const mockFields: FieldType[] = [
+  {
+    name: 'machine.os',
+    esTypes: ['text'],
+    type: 'string',
+    aggregatable: false,
+    searchable: false,
+    filterable: true,
+  },
+  {
+    name: 'machine.os.raw',
+    type: 'string',
+    esTypes: ['keyword'],
+    aggregatable: true,
+    searchable: true,
+    filterable: true,
+  },
+  {
+    name: 'not.filterable',
+    type: 'string',
+    esTypes: ['text'],
+    aggregatable: true,
+    searchable: false,
+    filterable: false,
+  },
+  {
+    name: 'bytes',
+    type: 'number',
+    esTypes: ['long'],
+    aggregatable: true,
+    searchable: true,
+    filterable: true,
+  },
+  {
+    name: '@timestamp',
+    type: 'date',
+    esTypes: ['date'],
+    aggregatable: true,
+    searchable: true,
+    filterable: true,
+  },
+  {
+    name: 'clientip',
+    type: 'ip',
+    esTypes: ['ip'],
+    aggregatable: true,
+    searchable: true,
+    filterable: true,
+  },
+  {
+    name: 'bool.field',
+    type: 'boolean',
+    esTypes: ['boolean'],
+    aggregatable: true,
+    searchable: true,
+    filterable: true,
+  },
+];
+
+export const mockIndexPattern: StaticIndexPattern = {
+  id: 'logstash-*',
+  fields: mockFields,
+  title: 'logstash-*',
+  timeFieldName: '@timestamp',
+};
