+++ src/legacy/ui/public/index_patterns/__mocks__/index.ts
@@ -46,11 +46,6 @@ export {
   IndexPatternMissingIndices,
   NoDefaultIndexPattern,
   NoDefinedIndexPatterns,
-} from '../../../../core_plugins/data/public';
-
-/* eslint-disable @kbn/eslint/no-restricted-paths */
-export {
   mockFields,
   mockIndexPattern,
-} from '../../../../core_plugins/data/public/index_patterns/index_patterns_service.mock';
-/* eslint-enable @kbn/eslint/no-restricted-paths */
+} from '../../../../core_plugins/data/public';
