+++ src/legacy/ui/public/index_patterns/index.ts
@@ -26,13 +26,6 @@
 
 import { setup as data } from '../../../core_plugins/data/public/legacy';
 
-/* eslint-disable @kbn/eslint/no-restricted-paths */
-export {
-  mockFields,
-  mockIndexPattern,
-} from '../../../core_plugins/data/public/index_patterns/index_patterns_service.mock';
-/* eslint-enable @kbn/eslint/no-restricted-paths */
-
 export const {
   FieldList, // only used in Discover and StubIndexPattern
   flattenHitWrapper,
@@ -55,6 +48,8 @@ export {
   IndexPatternMissingIndices,
   NoDefaultIndexPattern,
   NoDefinedIndexPatterns,
+  mockFields,
+  mockIndexPattern,
 } from '../../../core_plugins/data/public';
 
 // types
